excel done

+++ database/migrations/2022_12_13_061932_create_products_table.php
@@ -27,8 +27,8 @@ class CreateProductsTable extends Migration
             $table->string('image_three')->nullable();
             $table->string('image_four')->nullable();
             $table->string('image_five')->nullable();
-            $table->integer('cost')->nullable();
-            $table->integer('mrp_price');
+            $table->double('cost',8,2)->nullable();
+            $table->double('mrp_price',8,2);
             $table->string('dimension')->nullable();
             $table->string('country_of_origin')->nullable();
             $table->string('weight')->nullable();
